[core] Extract useEventCallback into its own module
