+++ packages/material-ui/src/ButtonBase/ButtonBase.js
@@ -4,6 +4,7 @@ import ReactDOM from 'react-dom';
 import clsx from 'clsx';
 import { elementTypeAcceptingRef } from '@material-ui/utils';
 import { useForkRef } from '../utils/reactHelpers';
+import useEventCallback from '../utils/useEventCallback';
 import withStyles from '../styles/withStyles';
 import NoSsr from '../NoSsr';
 import { useIsFocusVisible } from '../utils/focusVisible';
@@ -46,20 +47,6 @@ export const styles = {
   /* Pseudo-class applied to the root element if keyboard focused. */
   focusVisible: {},
 };
-
-const useEnhancedEffect = typeof window !== 'undefined' ? React.useLayoutEffect : React.useEffect;
-/**
- * https://github.com/facebook/react/issues/14099#issuecomment-440013892
- *
- * @param {function} fn
- */
-function useEventCallback(fn) {
-  const ref = React.useRef(fn);
-  useEnhancedEffect(() => {
-    ref.current = fn;
-  });
-  return React.useCallback(event => (0, ref.current)(event), []);
-}
 
 /**
  * `ButtonBase` contains as few styles as possible.
