+++ packages/material-ui/src/utils/index.js
@@ -6,4 +6,5 @@ export { default as ownerDocument } from './ownerDocument';
 export { default as ownerWindow } from './ownerWindow';
 export { default as requirePropFactory } from './requirePropFactory';
 export { default as unsupportedProp } from './unsupportedProp';
+export { default as useEventCallback } from './useEventCallback';
 export { default as withForwardedRef } from './withForwardedRef';
